more regexp, configuration, shortcuts, continue precedence
COMPATIBILITY NOTE
=======================================================================
Compatibility kept with older syntax. Although previous syntax allowed
only strings for prompt and comment, and only regexp for continue
prompt, this version allows regexp and strings everywhere.

To be compatible, if a string is given (old syntax) it is converted to
a regexp using a simple rule. every char become a class so all of them
loose their eventual special meaning and become themselves again. it is
really easy only the regexp delimiter (/) and \ need escaping in this
case. An example :

this string as a regexp would not do what the writer expects
"A[10] \*.?/rep> " ==>
    /[A][[][1][0][]][ ][\\][*][.][?][\/][r][e][p][>][ ]/

When converted, strings for prompts are prefixed with ^.*? to anchor
them at the begining of line and extand the matched part to the end
of the prompt. This is not done with the string for the comment. this
is done because we need the prompt length somewhere in the code.

	<cli prompt=">" comment=";">

	prompt_regexp =  /^.*?[>]/
	comment regexp = /[;]/
=======================================================================

Changes :
- nested cli now use a stack of prompt, so they work reliably.
- all params are ''_extract''ed from the <cli..>
- all three params may be strings or regexp
- configuration option for default prompts and comment
- configuration option to define named userdefined shortcuts
  for prompt and comments
- use the named shortcuts in <cli...>
- reordered the code analysing loop. Gives a higher precedence to
  continuation prompt
- lang english and french for the parametre.
- default settings should recognise most of unix cases.

/schplurtz